+++ pipeline-adapters/mlrun-pipelines-kfp-v1-8/src/mlrun_pipelines/ops.py
@@ -31,7 +31,7 @@ import mlrun.common.constants as mlrun_constants
 import mlrun.common.runtimes.constants
 import mlrun.utils.helpers
 from mlrun.config import config
-from mlrun.utils import get_in, logger
+from mlrun.utils import get_in
 
 # Disable the warning about reusing components
 dsl.ContainerOp._DISABLE_REUSABLE_COMPONENT_WARNING = True
@@ -257,24 +257,6 @@ def add_function_node_selection_attributes(
             container_op.affinity = function.spec.affinity
 
     return container_op
-
-
-def _enrich_node_selector_from_project(function):
-    function_node_selector = getattr(function.spec, "node_selector") or {}
-    project = mlrun.get_current_project()
-    if function.metadata.project != project.metadata.name:
-        project = function._get_db().get_project(function.metadata.project)
-    if project:
-        logger.debug(
-            "Enriching node selector from project",
-            project_node_selector=project.spec.default_function_node_selector,
-        )
-        return mlrun.utils.helpers.to_non_empty_values_dict(
-            mlrun.utils.helpers.merge_with_precedence(
-                project.spec.default_function_node_selector, function_node_selector
-            )
-        )
-    return mlrun.utils.helpers.to_non_empty_values_dict(function_node_selector)
 
 
 def generate_kfp_dag_and_resolve_project(run, project=None):
