+++ pipeline-adapters/mlrun-pipelines-kfp-v2/src/mlrun_pipelines/ops.py
@@ -15,6 +15,7 @@
 
 import os
 
+import mlrun_pipelines.common
 from kfp import dsl
 from kfp import kubernetes as kfp_k8s
 from mlrun_pipelines.common.helpers import (
@@ -110,7 +111,9 @@ def add_function_node_selection_attributes(
     function, task: dsl.PipelineTask
 ) -> dsl.PipelineTask:
     if not mlrun.runtimes.RuntimeKinds.is_local_runtime(function.kind):
-        enriched_node_selector = _enrich_node_selector_from_project(function)
+        enriched_node_selector = (
+            mlrun_pipelines.common.ops._enrich_node_selector_from_project(function)
+        )
         if enriched_node_selector:
             for k, v in enriched_node_selector.items():
                 task = kfp_k8s.add_node_selector(task, k, v)
@@ -143,24 +146,6 @@ def add_function_node_selection_attributes(
         )
 
     return task
-
-
-def _enrich_node_selector_from_project(function):
-    function_node_selector = getattr(function.spec, "node_selector") or {}
-    project = mlrun.get_current_project()
-    if function.metadata.project != project.metadata.name:
-        project = function._get_db().get_project(function.metadata.project)
-    if project:
-        logger.debug(
-            "Enriching node selector from project",
-            project_node_selector=project.spec.default_function_node_selector,
-        )
-        return mlrun.utils.helpers.to_non_empty_values_dict(
-            mlrun.utils.helpers.merge_with_precedence(
-                project.spec.default_function_node_selector, function_node_selector
-            )
-        )
-    return mlrun.utils.helpers.to_non_empty_values_dict(function_node_selector)
 
 
 def add_annotations(
